# specs/009-investment-workbench/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lucking 投资工作台 API
  version: 1.0.0
  description: 前端唯一公共契约；所有端点均为同域 /api/v1 路径。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /watchlists:
    get:
      summary: 获取当前用户自选分组及成员
      operationId: listWatchlists
      responses:
        '200':
          description: 自选分组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: 创建当前用户自选分组
      operationId: createWatchlist
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WatchlistGroupInput' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchlistGroupResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /watchlists/order:
    put:
      summary: 调整当前用户全部自选分组顺序
      operationId: orderWatchlists
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group_ids]
              properties:
                group_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: 排序后的自选分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchlistListResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /watchlists/{group_id}:
    put:
      summary: 修改当前用户自选分组
      operationId: updateWatchlist
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WatchlistGroupInput' }
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchlistGroupResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: 删除当前用户自选分组及成员
      operationId: deleteWatchlist
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: 删除成功 }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /watchlists/{group_id}/members:
    post:
      summary: 向当前用户自选分组添加股票
      operationId: addWatchlistMember
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stock_id]
              properties:
                stock_id: { type: string, format: uuid }
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchlistMemberResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /watchlists/{group_id}/members/{stock_id}:
    delete:
      summary: 从当前用户自选分组移除股票
      operationId: removeWatchlistMember
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/StockId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: 移除成功 }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /task-status:
    get:
      summary: 获取指定业务日期的实时归一任务状态
      operationId: listTaskStatus
      parameters:
        - name: business_date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: 任务状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskStatusApiResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /task-summaries/{business_date}:
    get:
      summary: 获取 20:00 持久化任务汇总快照
      operationId: getTaskSummary
      parameters:
        - name: business_date
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: 汇总快照及通知状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSummaryApiResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lucking_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
    GroupId:
      name: group_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    StockId:
      name: stock_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: 请求字段或业务范围无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BadRequestErrorResponse' }
    Unauthorized:
      description: 未认证或会话已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UnauthorizedErrorResponse' }
    Forbidden:
      description: CSRF、同源或资源所有权校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ForbiddenErrorResponse' }
    NotFound:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/NotFoundErrorResponse' }
    Conflict:
      description: 唯一性或并发版本冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ConflictErrorResponse' }
...
    InternalServerError:
      description: 未预期服务错误；不得返回堆栈、SQL、秘密或第三方原始响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/InternalServerErrorResponse' }
  schemas:
    BusinessErrorCode:
      type: integer
      description: 稳定业务错误码；HTTP 状态表达协议结果，业务码表达应用原因
      enum:
        - 100001
        - 100002
        - 100003
        - 200001
        - 200002
        - 200003
        - 300001
        - 400001
        - 400002
        - 400003
        - 400004
        - 500001
        - 500002
        - 900001
      x-enum-varnames:
        - INVALID_CREDENTIALS
        - SESSION_INVALID
        - CSRF_VALIDATION_FAILED
        - REQUEST_VALIDATION_FAILED
        - QUERY_RANGE_INVALID
        - PASSWORD_POLICY_VIOLATION
        - RESOURCE_NOT_FOUND
        - IMPORTANT_DATE_CONFLICT
        - WATCHLIST_NAME_CONFLICT
        - WATCHLIST_MEMBER_CONFLICT
        - WATCHLIST_CAPACITY_EXCEEDED
        - RATE_LIMITED
        - EXTERNAL_DEPENDENCY_UNAVAILABLE
        - INTERNAL_SERVER_ERROR
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: [string, 'null'] }
        code: { type: string }
        message: { type: string }
...
    SuccessResponseBase:
      type: object
      required: [code, message, errors, request_id, timestamp]
      properties:
        code: { type: integer, const: 0 }
        message: { type: string, const: '' }
        errors:
          type: array
          maxItems: 0
          items: { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, minLength: 1 }
        timestamp: { type: string, format: date-time }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, data, errors, request_id, timestamp]
      properties:
        code: { $ref: '#/components/schemas/BusinessErrorCode' }
        message: { type: string, minLength: 1 }
        data: { type: 'null' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, minLength: 1 }
        timestamp: { type: string, format: date-time }
    BadRequestErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [200001, 200002, 200003] }
    UnauthorizedErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [100001, 100002] }
    ForbiddenErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 100003 }
    NotFoundErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 300001 }
    ConflictErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [400001, 400002, 400003, 400004] }
...
    InternalServerErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 900001 }
...
    StockSummary:
      type: object
      required: [stock_id, market_code, venue_code, security_code, name, listing_status]
      properties:
        stock_id: { type: string, format: uuid }
        market_code: { type: string }
        venue_code: { type: string }
        security_code: { type: string }
        name: { type: string }
        listing_status: { type: string }
...
    WatchlistGroupInput:
      type: object
      required: [name, notes, tags]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        notes: { type: string, minLength: 1, maxLength: 1000 }
        tags:
          type: array
          minItems: 1
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 30 }
    WatchlistMember:
      type: object
      required: [member_id, stock, sort_order]
      properties:
        member_id: { type: string, format: uuid }
        stock: { $ref: '#/components/schemas/StockSummary' }
        sort_order: { type: integer, minimum: 0 }
    WatchlistGroup:
      allOf:
        - $ref: '#/components/schemas/WatchlistGroupInput'
        - type: object
          required: [group_id, sort_order, members]
          properties:
            group_id: { type: string, format: uuid }
            sort_order: { type: integer, minimum: 0 }
            members:
              type: array
              items: { $ref: '#/components/schemas/WatchlistMember' }
...
    TaskStatusItem:
      type: object
      required: [task_key, schedule_slug, display_name, status, observed_at]
      properties:
        task_key: { type: string }
        schedule_slug: { type: string }
        display_name: { type: string }
        status: { type: string, enum: [SUCCEEDED, PARTIAL, FAILED, RUNNING, UNKNOWN, NOT_RUN] }
        source_run_id: { type: [string, 'null'] }
        started_at: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        record_count: { type: [integer, 'null'], minimum: 0 }
        error_category: { type: [string, 'null'] }
        error_summary: { type: [string, 'null'] }
        observed_at: { type: string, format: date-time }
    TaskStatusResponse:
      type: object
      required: [business_date, observed_at, items]
      properties:
        business_date: { type: string, format: date }
        observed_at: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/TaskStatusItem' }
    TaskSummary:
      type: object
      required:
        - summary_id
        - business_date
        - status
        - notification_status
        - counts
        - items
        - latest_notification_attempt
      properties:
        summary_id: { type: string, format: uuid }
        business_date: { type: string, format: date }
        status: { type: string, enum: [BUILDING, READY, FAILED] }
        notification_status: { type: string, enum: [PENDING, SENDING, SENT, FAILED] }
        generated_at: { type: [string, 'null'], format: date-time }
        notified_at: { type: [string, 'null'], format: date-time }
        counts: { $ref: '#/components/schemas/TaskStatusCounts' }
        items:
          type: array
          items: { $ref: '#/components/schemas/TaskStatusItem' }
        latest_notification_attempt:
          oneOf: [{ $ref: '#/components/schemas/NotificationAttempt' }, { type: 'null' }]
    TaskStatusCounts:
      type: object
      additionalProperties: false
      required: [SUCCEEDED, PARTIAL, FAILED, RUNNING, UNKNOWN, NOT_RUN]
      properties:
        SUCCEEDED: { type: integer, minimum: 0 }
        PARTIAL: { type: integer, minimum: 0 }
        FAILED: { type: integer, minimum: 0 }
        RUNNING: { type: integer, minimum: 0 }
        UNKNOWN: { type: integer, minimum: 0 }
        NOT_RUN: { type: integer, minimum: 0 }
    NotificationAttempt:
      type: object
      additionalProperties: false
      required:
        - attempt_no
        - trigger_kind
        - status
        - error_category
        - error_summary
        - started_at
        - completed_at
        - retryable
      properties:
        attempt_no: { type: integer, minimum: 1 }
        trigger_kind: { type: string, enum: [AUTOMATIC, MANUAL_RETRY] }
        status: { type: string, enum: [RUNNING, SUCCEEDED, FAILED] }
        error_category: { type: [string, 'null'] }
        error_summary: { type: [string, 'null'] }
        started_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        retryable: { type: boolean }
...
    WatchlistListResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/WatchlistGroup' }
    WatchlistGroupResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/WatchlistGroup' }
    WatchlistMemberResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/WatchlistMember' }
...
    TaskStatusApiResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/TaskStatusResponse' }
    TaskSummaryApiResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/TaskSummary' }
...